Complete edit workflow modular refactoring
- Move src/ui/edit.rs functions to src/ui/edit_workflow/ui.rs
- Create pure edit workflow functions returning Element<EditMessage>
- Add missing EditMessage variants for navigation (BackToMainMenu, EditAnother)
- Update edit_workflow module to include UI exports
- Create shared configuration UI module at src/ui/shared/configuration.rs
- Remove old edit.rs file and update module imports
- All edit workflow functions now follow Iced scaling pattern
- Edit workflow handlers map to app-level messages when needed

Both flash and edit workflows now properly implement modular architecture
