+++ src/ui.rs
@@ -1,5 +1,4 @@
 pub mod application;
-pub mod edit;
 mod icons;
 pub mod preset_editor;
 pub mod start_screen;
@@ -11,17 +10,22 @@ pub mod preset_manager;
 pub mod device_selection;
 pub mod configuration;
 
+// Shared UI components
+pub mod shared;
+
 // Unified message system
 pub mod messages;
 
 #[allow(unused_imports)]
 pub use application::GolemGpuImager;
-pub use edit::{view_edit_completion, view_edit_configuration, view_select_existing_device};
+pub use edit_workflow::{view_edit_completion, view_edit_configuration, view_select_existing_device};
 #[allow(unused_imports)]
 pub use flash_workflow::{
-    view_configure_settings, view_flash_completion, view_preset_editor, view_select_os_image,
+    view_flash_completion, view_select_os_image,
     view_select_target_device, view_writing_process,
 };
+#[allow(unused_imports)]
+pub use shared::{view_configuration_editor, view_preset_editor};
 #[allow(unused_imports)]
 use iced::Font;
 pub use start_screen::view_start_screen;
