+++ src/ui/edit_workflow.rs
@@ -1,7 +1,9 @@
 pub mod state;
 pub mod messages;
 pub mod handler;
+pub mod ui;
 
 pub use state::*;
 pub use messages::*;
 pub use handler::*;
+pub use ui::*;
