+++ src/ui/flash_workflow/ui.rs
@@ -1,16 +1,12 @@
 use iced::alignment::Horizontal;
 use iced::widget::{
-    Column, Container, button, column, container, pick_list, progress_bar, row, scrollable, svg,
+    Column, Container, button, column, container, progress_bar, row, scrollable, svg,
     text,
 };
 use iced::{Alignment, Color, Element, Length};
 use iced::{Border, Theme};
-
-use crate::models::{NetworkType, PaymentNetwork};
 use crate::ui::{
-    messages::Message,
     device_selection::StorageDevice,
-    preset_manager::PresetEditorMessage,
 };
 use super::{FlashMessage, OsImage, OsImageGroup};
 use crate::style;
@@ -746,575 +742,6 @@ pub fn view_processing_image(
         .width(Length::Fill)
         .height(Length::Fill)
         .center_x(Length::Fill)
-        .into()
-}
-
-/// Shared configuration UI component used in both flash and edit workflows
-pub fn view_configuration_editor<'a>(
-    payment_network: PaymentNetwork,
-    subnet: String,
-    network_type: NetworkType,
-    wallet_address: String,
-    is_wallet_valid: bool,
-    title_text: &'a str,
-    description_text: &'a str,
-    back_action: Message,
-    next_action: Message,
-    back_label: &'a str,
-    next_label: &'a str,
-    configuration_presets: &'a [crate::models::ConfigurationPreset],
-    selected_preset: Option<usize>,
-    new_preset_name: &'a str,
-    show_preset_manager: bool,
-    preset_editor: Option<&'a crate::ui::preset_manager::PresetEditor>,
-    preset_back_action: Message,
-) -> Element<'a, Message> {
-    // Create the title - simple text by default
-    let title = text(title_text).size(30);
-
-    // Create simplified settings UI
-    let description = text(description_text).size(16);
-
-    // Network selection
-    let network_label = text("Payment Network").size(18);
-
-    let network_pl = pick_list(
-        &[PaymentNetwork::Testnet, PaymentNetwork::Mainnet][..],
-        Some(payment_network),
-        FlashMessage::SetPaymentNetwork,
-    )
-    .style(crate::style::pick_list_style);
-
-    // Subnet field with text input
-    let subnet_label = text("Subnet").size(18);
-    let subnet_input = iced::widget::text_input("Enter subnet", &subnet)
-        .on_input(FlashMessage::SetSubnet)
-        .padding(10);
-
-    // Wallet address field with text input and validation indicator
-    let wallet_label = text("Ethereum Wallet Address").size(18);
-
-    // Choose style based on validation state and create input with appropriate styling
-    let wallet_input = if !wallet_address.is_empty() {
-        if is_wallet_valid {
-            iced::widget::text_input("Enter ETH wallet address", &wallet_address)
-                .on_input(FlashMessage::SetWalletAddress)
-                .padding(10)
-                .style(crate::style::valid_wallet_input)
-        } else {
-            iced::widget::text_input("Enter ETH wallet address", &wallet_address)
-                .on_input(FlashMessage::SetWalletAddress)
-                .padding(10)
-                .style(crate::style::invalid_wallet_input)
-        }
-    } else {
-        // Default styling for empty input
-        iced::widget::text_input("Enter ETH wallet address", &wallet_address)
-            .on_input(FlashMessage::SetWalletAddress)
-            .padding(10)
-            .style(crate::style::default_text_input)
-    };
-
-    // Add validation message if address is not empty
-    let validation_message = if !wallet_address.is_empty() {
-        if is_wallet_valid {
-            container(row![icons::checkmark(), text("Valid Ethereum address").size(14)].spacing(5))
-                .style(crate::style::valid_message_container)
-        } else {
-            container(
-                row![
-                    icons::error(),
-                    text("Invalid Ethereum address format").size(14)
-                ]
-                .spacing(5),
-            )
-            .style(crate::style::invalid_message_container)
-        }
-    } else {
-        container(row![text("").size(14)])
-    };
-
-    // Network type using pick list
-    let type_label = text("Network Type").size(18);
-
-    let type_pl = pick_list(
-        &[NetworkType::Hybrid, NetworkType::Central][..],
-        Some(network_type),
-        FlashMessage::SetNetworkType,
-    )
-    .style(crate::style::pick_list_style);
-
-    // Navigation buttons
-    let back_button = button(
-        row![icons::navigate_before(), text(back_label)]
-            .spacing(5)
-            .align_y(Alignment::Center),
-    )
-    .on_press(back_action)
-    .padding(10)
-    .style(button::secondary);
-
-    // Use different icons based on the label
-    let next_icon = if next_label.contains("Save") {
-        icons::save()
-    } else {
-        icons::send()
-    };
-
-    let next_button = button(
-        row![text(next_label), next_icon]
-            .spacing(5)
-            .align_y(Alignment::Center),
-    )
-    .on_press(next_action)
-    .padding(10)
-    .style(button::primary);
-
-    let navigation = row![back_button, next_button]
-        .spacing(15)
-        .width(Length::Fill)
-        .align_y(Alignment::Center);
-
-    // We'll create all UI components inline for better type inference
-
-    // Main configuration content
-    let config_content = column![
-        network_label,
-        network_pl.width(Length::Fill),
-        subnet_label,
-        subnet_input,
-        wallet_label,
-        wallet_input,
-        validation_message,
-        type_label,
-        type_pl.width(Length::Fill),
-    ]
-    .spacing(10)
-    .width(Length::Fill);
-
-    // Check if we should show preset editor, preset manager, or normal configuration UI
-    if let Some(editor) = preset_editor {
-        // Show preset editor
-        return view_preset_editor(editor);
-    } else if show_preset_manager {
-        // Create a preset management UI that allows detailed management of presets
-        let preset_list = if configuration_presets.is_empty() {
-            container(text("No presets defined. Create your first preset.").size(16))
-                .width(Length::Fill)
-                .padding(20)
-                .center_x(Length::Fill)
-        } else {
-            let mut preset_rows = column![];
-
-            // Create an entry for each preset with full management options
-            for (idx, preset) in configuration_presets.iter().enumerate() {
-                let is_selected = selected_preset == Some(idx);
-                let is_default = preset.is_default;
-
-                // Create a row for each preset with name, actions, and info
-                let preset_row = container(
-                    row![
-                        // Star icon for default preset
-                        if is_default {
-                            icons::star()
-                        } else {
-                            icons::star_border()
-                        },
-                        // Preset name
-                        text(&preset.name).size(16).width(Length::Fill),
-                        // Network info
-                        column![
-                            text(format!("Network: {:?}", preset.payment_network)).size(14),
-                            text(format!("Type: {:?}", preset.network_type)).size(14),
-                        ]
-                        .width(Length::Fill),
-                        // Action buttons
-                        button(
-                            row![icons::delete(), text("Delete")]
-                                .spacing(5)
-                                .align_y(Alignment::Center)
-                        )
-                        .on_press(Message::DeletePreset(idx))
-                        .style(button::danger)
-                        .padding(8),
-                        button(
-                            row![icons::star(), text("Set Default")]
-                                .spacing(5)
-                                .align_y(Alignment::Center)
-                        )
-                        .on_press(Message::SetDefaultPreset(idx))
-                        .style(if is_default {
-                            button::success
-                        } else {
-                            button::secondary
-                        })
-                        .padding(8),
-                        button(
-                            row![icons::edit(), text("Edit")]
-                                .spacing(5)
-                                .align_y(Alignment::Center)
-                        )
-                        .on_press(Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::Start(idx))))
-                        .style(button::secondary)
-                        .padding(8),
-                        button(
-                            row![icons::tune(), text("Load")]
-                                .spacing(5)
-                                .align_y(Alignment::Center)
-                        )
-                        .on_press(Message::SelectPreset(idx))
-                        .style(if is_selected {
-                            button::primary
-                        } else {
-                            button::secondary
-                        })
-                        .padding(8)
-                    ]
-                    .spacing(10)
-                    .align_y(Alignment::Center),
-                )
-                .padding(10)
-                .style(if is_selected {
-                    crate::style::selected_container
-                } else {
-                    crate::style::bordered_box
-                })
-                .width(Length::Fill);
-
-                preset_rows = column![preset_rows, preset_row];
-            }
-
-            container(
-                column![
-                    row![
-                        icons::settings(),
-                        text("Manage Configuration Presets").size(20)
-                    ]
-                    .spacing(5)
-                    .align_y(Alignment::Center),
-                    container(preset_rows)
-                        .padding(10)
-                        .style(crate::style::bordered_box)
-                ]
-                .spacing(10),
-            )
-            .width(Length::Fill)
-        };
-
-        // Add new preset form
-        let new_preset_form = container(
-            column![
-                text("Create New Preset").size(18),
-                row![
-                    iced::widget::text_input("Enter preset name", new_preset_name)
-                        .on_input(Message::SetPresetName)
-                        .padding(10)
-                        .width(Length::Fill),
-                    button(
-                        row![icons::save(), text("Save")]
-                            .spacing(5)
-                            .align_y(Alignment::Center)
-                    )
-                    .on_press(Message::SaveAsPreset)
-                    .style(button::primary)
-                    .padding(10),
-                ]
-                .spacing(10)
-            ]
-            .spacing(10),
-        )
-        .padding(15)
-        .style(crate::style::bordered_box)
-        .width(Length::Fill);
-
-        // Return button
-        let back_to_config = button(
-            row![icons::navigate_before(), text("Back to Configuration")]
-                .spacing(8)
-                .align_y(Alignment::Center),
-        )
-        .on_press(preset_back_action)
-        .padding(10)
-        .style(button::secondary);
-
-        // Layout for preset manager view
-        column![
-            title,
-            text("Manage your saved configuration presets").size(16),
-            preset_list,
-            new_preset_form,
-            back_to_config
-        ]
-        .spacing(20)
-        .padding(20)
-        .into()
-    } else {
-        // Main layout with clean vertical organization (normal view)
-        column![
-            title,
-            description,
-            // Preset picker at the top using pick_list
-            container(
-                row![
-                    text("Preset:").size(16),
-                    if !configuration_presets.is_empty() {
-                        // Get the currently selected preset, if any
-                        let selected = selected_preset.and_then(|idx| {
-                            if idx < configuration_presets.len() {
-                                Some(&configuration_presets[idx])
-                            } else {
-                                None
-                            }
-                        });
-
-                        // Create a pick_list for selecting presets
-                        let preset_picker = pick_list(configuration_presets, selected, |preset| {
-                            // Find the index of the selected preset
-                            let idx = configuration_presets
-                                .iter()
-                                .position(|p| p.name == preset.name)
-                                .unwrap_or(0);
-
-                            Message::SelectPreset(idx)
-                        })
-                        .width(Length::Fill)
-                        .style(crate::style::pick_list_style);
-
-                        let row_with_icon = row![icons::tune(), preset_picker]
-                            .spacing(5)
-                            .align_y(Alignment::Center)
-                            .width(Length::Fill);
-
-                        let preset_container: Element<'_, Message> =
-                            container(row_with_icon).width(Length::Fill).into();
-
-                        preset_container
-                    } else {
-                        // Show a disabled text input when no presets are available
-                        container(text("No presets available").size(16))
-                            .width(Length::Fill)
-                            .padding(8)
-                            .style(crate::style::bordered_box)
-                            .into()
-                    },
-                    button(
-                        row![icons::settings(), text("Manage").size(14)]
-                            .spacing(5)
-                            .align_y(Alignment::Center)
-                    )
-                    .on_press(Message::TogglePresetManager)
-                    .padding(8)
-                    .style(button::secondary)
-                ]
-                .spacing(10)
-                .align_y(Alignment::Center)
-            )
-            .padding(10)
-            .style(crate::style::bordered_box)
-            .width(Length::Fill),
-            // Main configuration content
-            config_content,
-            // Save as preset UI
-            container(
-                row![
-                    icons::save(),
-                    text("Save Current Settings as Preset").size(16),
-                    iced::widget::text_input("Enter preset name", new_preset_name)
-                        .on_input(Message::SetPresetName)
-                        .padding(8)
-                        .width(Length::Fill),
-                    button("Save")
-                        .on_press(Message::SaveAsPreset)
-                        .style(button::primary)
-                        .padding(8)
-                ]
-                .spacing(10)
-                .align_y(Alignment::Center)
-            )
-            .padding(10)
-            .style(crate::style::bordered_box)
-            .width(Length::Fill),
-            // Navigation buttons
-            navigation
-        ]
-        .spacing(20)
-        .padding(20)
-        .into()
-    }
-}
-
-pub fn view_configure_settings<'a>(
-    payment_network: PaymentNetwork,
-    subnet: String,
-    network_type: NetworkType,
-    wallet_address: String,
-    is_wallet_valid: bool,
-    configuration_presets: &'a [crate::models::ConfigurationPreset],
-    selected_preset: Option<usize>,
-    new_preset_name: &'a str,
-    show_preset_manager: bool,
-    preset_editor: Option<&'a crate::ui::preset_manager::PresetEditor>,
-) -> Element<'a, Message> {
-    view_configuration_editor(
-        payment_network,
-        subnet,
-        network_type,
-        wallet_address,
-        is_wallet_valid,
-        "Configure OS Image",
-        "Configure the OS image with the following options:",
-        Message::BackToSelectOsImage,
-        Message::WriteImage,
-        "Back",
-        "Write Image to Device",
-        configuration_presets,
-        selected_preset,
-        new_preset_name,
-        show_preset_manager,
-        preset_editor,
-        Message::TogglePresetManager,
-    )
-}
-
-pub fn view_preset_editor<'a>(
-    editor: &'a crate::ui::preset_manager::PresetEditor,
-) -> Element<'a, Message> {
-    use crate::models::{NetworkType, PaymentNetwork};
-    use crate::ui::preset_manager::PresetEditorMessage;
-    use crate::ui::icons;
-    use iced::widget::{button, column, container, pick_list, row, text, text_input};
-    use iced::{Alignment, Length};
-
-    // Calculate wallet validation based on the current editor state
-    let is_wallet_valid = editor.wallet_address.is_empty() || 
-        crate::utils::eth::is_valid_eth_address(&editor.wallet_address);
-
-    let title = text(format!("Edit Preset: {}", editor.name)).size(24);
-
-    let name_input = container(
-        column![
-            text("Preset Name").size(16),
-            text_input("Enter preset name", &editor.name)
-                .on_input(|name| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateName(name))))
-                .padding(10)
-                .width(Length::Fill)
-        ]
-        .spacing(5)
-    )
-    .width(Length::Fill);
-
-    let network_selection = container(
-        column![
-            text("Payment Network").size(16),
-            pick_list(
-                &[PaymentNetwork::Testnet, PaymentNetwork::Mainnet][..],
-                Some(editor.payment_network),
-                |network| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdatePaymentNetwork(network))),
-            )
-            .style(crate::style::pick_list_style)
-        ]
-        .spacing(5)
-    )
-    .width(Length::Fill);
-
-    let subnet_input = container(
-        column![
-            text("Subnet").size(16),
-            text_input("Enter subnet", &editor.subnet)
-                .on_input(|subnet| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateSubnet(subnet))))
-                .padding(10)
-                .width(Length::Fill)
-        ]
-        .spacing(5)
-    )
-    .width(Length::Fill);
-
-    let network_type_selection = container(
-        column![
-            text("Network Type").size(16),
-            pick_list(
-                &[NetworkType::Central, NetworkType::Hybrid][..],
-                Some(editor.network_type),
-                |network_type| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateNetworkType(network_type))),
-            )
-            .style(crate::style::pick_list_style)
-        ]
-        .spacing(5)
-    )
-    .width(Length::Fill);
-
-    let wallet_input = container(
-        column![
-            text("Ethereum Wallet Address").size(16),
-            if !editor.wallet_address.is_empty() {
-                if is_wallet_valid {
-                    text_input("Enter ETH wallet address", &editor.wallet_address)
-                        .on_input(|addr| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateWalletAddress(addr))))
-                        .padding(10)
-                        .style(crate::style::valid_wallet_input)
-                } else {
-                    text_input("Enter ETH wallet address", &editor.wallet_address)
-                        .on_input(|addr| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateWalletAddress(addr))))
-                        .padding(10)
-                        .style(crate::style::invalid_wallet_input)
-                }
-            } else {
-                text_input("Enter ETH wallet address", &editor.wallet_address)
-                    .on_input(|addr| Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::UpdateWalletAddress(addr))))
-                    .padding(10)
-            }
-        ]
-        .spacing(5)
-    )
-    .width(Length::Fill);
-
-    let form = container(
-        column![
-            name_input,
-            network_selection,
-            subnet_input,
-            network_type_selection,
-            wallet_input,
-        ]
-        .spacing(15)
-    )
-    .padding(20)
-    .style(crate::style::bordered_box)
-    .width(Length::Fill);
-
-    let actions = row![
-        button(
-            row![icons::navigate_before(), text("Cancel")]
-                .spacing(8)
-                .align_y(Alignment::Center)
-        )
-        .on_press(Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::Cancel)))
-        .style(button::secondary)
-        .padding(10),
-        
-        button(
-            row![icons::save(), text("Save Changes")]
-                .spacing(8)
-                .align_y(Alignment::Center)
-        )
-        .on_press_maybe(if editor.is_valid() {
-            Some(Message::PresetManager(crate::ui::preset_manager::PresetManagerMessage::Editor(PresetEditorMessage::Save)))
-        } else {
-            None
-        })
-        .style(if editor.is_valid() {
-            button::primary
-        } else {
-            button::secondary
-        })
-        .padding(10)
-    ]
-    .spacing(10)
-    .width(Length::Fill);
-
-    column![title, form, actions]
-        .spacing(20)
-        .padding(20)
         .into()
 }
