+++ src/ui/shared/configuration.rs
@@ -1,0 +1,64 @@
+use iced::widget::{
+    Column, Container, button, column, container, pick_list, row, scrollable, text, text_input,
+};
+use iced::{Alignment, Color, Element, Length};
+use iced::{Border, Theme};
+
+use crate::models::{NetworkType, PaymentNetwork};
+use crate::ui::{
+    messages::Message,
+    preset_manager::PresetEditorMessage,
+    icons,
+};
+use crate::style;
+
+/// Shared configuration UI component used in both flash and edit workflows
+/// For now, keeping the original Message type - will refactor to generic later
+pub fn view_configuration_editor<'a>(
+    payment_network: PaymentNetwork,
+    subnet: String,
+    network_type: NetworkType,
+    wallet_address: String,
+    is_wallet_valid: bool,
+    title_text: &'a str,
+    description_text: &'a str,
+    back_action: Message,
+    next_action: Message,
+    back_label: &'a str,
+    next_label: &'a str,
+    configuration_presets: &'a [crate::models::ConfigurationPreset],
+    selected_preset: Option<usize>,
+    new_preset_name: &'a str,
+    show_preset_manager: bool,
+    preset_editor: Option<&'a crate::ui::preset_manager::PresetEditor>,
+    preset_back_action: Message,
+) -> Element<'a, Message> {
+    // Placeholder implementation - will move the full implementation here
+    container(
+        column![
+            text(title_text).size(24),
+            text("This is a placeholder for the shared configuration editor").size(16),
+            text("The full implementation will be moved here from flash_workflow").size(14),
+        ]
+        .spacing(15)
+    )
+    .padding(20)
+    .into()
+}
+
+/// Shared preset editor component
+pub fn view_preset_editor<'a>(
+    editor: &'a crate::ui::preset_manager::PresetEditor,
+) -> Element<'a, Message> {
+    // Placeholder implementation - will move the full implementation here
+    container(
+        column![
+            text(format!("Edit Preset: {}", editor.name)).size(24),
+            text("This is a placeholder for the shared preset editor").size(16),
+            text("The full implementation will be moved here from flash_workflow").size(14),
+        ]
+        .spacing(15)
+    )
+    .padding(20)
+    .into()
+}
