+++ src/ui/shared/mod.rs
@@ -1,0 +1,3 @@
+pub mod configuration;
+
+pub use configuration::*;
